// src/app/enums/pokemon-types.enum.ts
export enum Types {
  fire = 'fire',
  grass = 'grass',
  water = 'water',
  electrik = 'electrik',
  fly = 'fly',
  rock = 'rock',
  spectre = 'spectre',
  darkness = 'darkness',
  fight = 'fight',
  steel = 'steel',
  dragon = 'dragon',
  insect = 'insect',
  normal = 'normal'
}

export enum Gender {
  male = 'male',
  female = 'female'
}

export enum Color {
  fireColor = '#f75231',
  grassColor = '#7bce52',
  waterColor = '#399cff',
  electrikColor = '#ffc631',
  flyColor = '#ada594',
  rockColor = '#bda55a',
  spectreColor = '#6363b5',
  darknessColor = '#735a4a',
  fightColor = '#a55239',
  steelColor = '#ada594',
  dragonColor = '#8858f6',
  insectColor = '#adbd21',
  normalColor = '#ada594',
  fairyColor = '#e09ae3',
  psyColor = '#ff73a5',
  groundColor = '#d6b55a',
  poisonColor = '#b55aa5'
}